requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Utilities
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Urdu Text Processing
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# Logging and Monitoring
tensorboard>=2.13.0
